# Re-implement NLog logs injection