--- a/tracer/src/Datadog.Trace/ClrProfiler/ClrNames.cs
+++ b/tracer/src/Datadog.Trace/ClrProfiler/ClrNames.cs
@@ -45,5 +45,7 @@ namespace Datadog.Trace.ClrProfiler
         public const string GenericParameterTask = "System.Threading.Tasks.Task`1<T>";
         public const string ObjectTask = "System.Threading.Tasks.Task`1<System.Object>";
         public const string Int32Task = "System.Threading.Tasks.Task`1<System.Int32>";
+
+        public const string Type = "System.Type";
     }
 }