--- a/tracer/test/Datadog.Trace.ClrProfiler.IntegrationTests/NLogTests.cs
+++ b/tracer/test/Datadog.Trace.ClrProfiler.IntegrationTests/NLogTests.cs
@@ -12,52 +12,14 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests
 {
     public class NLogTests : LogsInjectionTestBase
     {
-        private readonly LogFileTest[] _nlogPre40LogFileTests =
-            {
-                new LogFileTest()
-                {
-                    FileName = "log-textFile.log",
-                    RegexFormat = @"{0}: {1}",
-                    UnTracedLogTypes = UnTracedLogTypes.EmptyProperties,
-                    PropertiesUseSerilogNaming = false
-                }
-            };
-
-        private readonly LogFileTest[] _nlog40LogFileTests =
-            {
-                new LogFileTest()
-                {
-                    FileName = "log-textFile.log",
-                    RegexFormat = @"{0}: {1}",
-                    UnTracedLogTypes = UnTracedLogTypes.EmptyProperties,
-                    PropertiesUseSerilogNaming = false
-                },
-                new LogFileTest()
-                {
-                    FileName = "log-jsonFile.log",
-                    RegexFormat = @"""{0}"": {1}",
-                    UnTracedLogTypes = UnTracedLogTypes.EmptyProperties,
-                    PropertiesUseSerilogNaming = false
-                }
-            };
-
-        private readonly LogFileTest[] _nlogPost40LogFileTests =
-            {
-                new LogFileTest()
-                {
-                    FileName = "log-textFile.log",
-                    RegexFormat = @"{0}: {1}",
-                    UnTracedLogTypes = UnTracedLogTypes.EmptyProperties,
-                    PropertiesUseSerilogNaming = false
-                },
-                new LogFileTest()
-                {
-                    FileName = "log-jsonFile.log",
-                    RegexFormat = @"""{0}"": {1}",
-                    UnTracedLogTypes = UnTracedLogTypes.None,
-                    PropertiesUseSerilogNaming = false
-                }
-            };
+        private readonly LogFileTest _textFile = new()
+        {
+            FileName = "log-textFile.log",
+            RegexFormat = @"{0}: {1}",
+            // txt format can't conditionally add properties
+            UnTracedLogTypes = UnTracedLogTypes.EmptyProperties,
+            PropertiesUseSerilogNaming = false
+        };
 
         public NLogTests(ITestOutputHelper output)
             : base(output, "LogsInjection.NLog")
@@ -98,17 +60,38 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests
                 var spans = agent.WaitForSpans(1, 2500);
                 Assert.True(spans.Count >= 1, $"Expecting at least 1 span, only received {spans.Count}");
 
-                var testFiles = GetTestFiles(packageVersion);
+                var testFiles = GetTestFiles(packageVersion, logsInjectionEnabled: false);
                 ValidateLogCorrelation(spans, testFiles, packageVersion, disableLogCorrelation: true);
             }
         }
 
-        private LogFileTest[] GetTestFiles(string packageVersion)
-            => packageVersion switch
+        private LogFileTest[] GetTestFiles(string packageVersion, bool logsInjectionEnabled = true)
+        {
+            if (packageVersion is null or "")
             {
-                null or "" => _nlogPre40LogFileTests,
-                _ when new Version(packageVersion) >= new Version("4.1.0") => _nlogPost40LogFileTests,
-                _ => _nlog40LogFileTests,
+                // pre 4.0 can't write to text file
+                return new[] { _textFile };
+            }
+
+            var unTracedLogType = new Version(packageVersion) switch
+            {
+                // When logs injection is enabled, untraced logs get env, service etc
+                { } x when x >= new Version("4.1.0") && logsInjectionEnabled => UnTracedLogTypes.EnvServiceTracingPropertiesOnly,
+                // When logs injection is enabled, no enrichment
+                { } x when x >= new Version("4.1.0") && !logsInjectionEnabled => UnTracedLogTypes.None,
+                // older version of nlog always adds empty properties
+                _ => UnTracedLogTypes.EmptyProperties
             };
+
+            return new[] { _textFile, GetJsonTestFile(unTracedLogType) };
+        }
+
+        private LogFileTest GetJsonTestFile(UnTracedLogTypes unTracedLogType) => new()
+        {
+            FileName = "log-jsonFile.log",
+            RegexFormat = @"""{0}"": {1}",
+            UnTracedLogTypes = unTracedLogType,
+            PropertiesUseSerilogNaming = false
+        };
     }
 }