… tracer/src/Datadog.Trace/Logging/LogsInjection/CustomNLogLogProvider.cs
… tracer/src/Datadog.Trace/Logging/LogsInjection/FallbackNLogLogProvider.cs
… tracer/test/Datadog.Trace.ClrProfiler.Managed.Tests/Datadog.Trace.ClrProfiler.Managed.Tests.csproj
… tracer/test/Datadog.Trace.Tests/Logging/NLogLogProviderTests.cs
… tracer/test/benchmarks/Benchmarks.Trace/NLogBenchmark.cs
… tracer/test/test-applications/integrations/LogsInjection.NLog/Program.cs