## Changes

Similar to #1710, this change modifies NLog's logs-injection code to stop using the LibLog based subscriber (which is not safe to use across AppDomain boundaries for some versions of NLog) and to use the MDLC/MDC directly using automatic instrumentation.

This change has a big limitation: **NLog logs injection now requires automatic instrumentation**.

Additionally, the CustomNLogLogProvider type has been renamed to NoOpNLogProvider and modified so that the LibLog interface does not affect the Log4Net state.

> This PR was originally built on top of the direct-log-submission PR #1945, but I rebased it to be on top of master instead

## Performance Considerations

The new approach still has a significant overhead compared to no logs injection. For small numbers of log messages written per trace (<10), the new approach shows a small _improvement_ over the LibLog provider. For larger numbers of logs-per-trace, the LibLog provider is significantly faster. 

Also, worth bearing in mind, that if the `Info` level isn't enabled, we don't run with the new version, so this is a worst-case simulation (in which every log in a trace is written) compared to the lib log, (which always enriches for a trace, regardless of whether any actual logs are written or enabled).

To test the performance, I ran the following benchmark code:

### Test Code
```csharp
[Benchmark]
public void WriteLogs()
{
    using (LogInjectionTracer.StartActive("Test"))
    {
        using (LogInjectionTracer.StartActive("Child"))
        {
            for (int i = 0; i < N; i++)
            {
                Logger.Info("Hello");
            }
        }
    }
}
```

### Benchmark results

The following are for `netcoreapp3.1` only - I was having issues with Sentinel killing the process with `net472` 🤷 



|      Method |   N |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---- |-----------:|----------:|----------:|-------:|------:|------:|----------:|
| None |   1 |  3.520 us | 0.0675 us | 0.0750 us | 0.0458 |     - |     - |   2.91 KB |
| None |  10 | 11.963 us | 0.0544 us | 0.0482 us | 0.1678 |     - |     - |  11.07 KB |
| None | 100 | 91.977 us | 0.8500 us | 0.7535 us | 1.3428 |     - |     - |  92.63 KB |
|  |  | | | | | | | |
| Previous (LibLog)|   1 |   6.318 us | 0.0357 us | 0.0278 us | 0.0916 |     - |     - |   6.28 KB |
| Previous (LibLog)|  10 |  18.607 us | 0.0612 us | 0.0542 us | 0.2747 |     - |     - |  18.51 KB |
| Previous (LibLog)| 100 | 139.824 us | 1.5014 us | 1.3309 us | 1.9531 |     - |     - | 140.84 KB |
|  |  | | | | | | | |
| New (Auto-instrument)|   1 |   5.027 us | 0.0505 us | 0.0422 us | 0.0610 |     - |     - |    4.4 KB |
| New (Auto-instrument)|  10 |  25.568 us | 0.2867 us | 0.2682 us | 0.3662 |     - |     - |  24.86 KB |
| New (Auto-instrument)| 100 | 223.142 us | 0.7888 us | 0.7379 us | 3.4180 |     - |     - | 229.47 KB |

@DataDog/apm-dotnet